拡張メソッドの定義を分離
