+++ Assets/Scripts/CerkeOnline/Application/UseCaseFactory/SetPieceUseCaseFactory.cs
@@ -15,24 +15,4 @@ namespace Azarashi.CerkeOnline.Application
             return new PlayerSetPieceUseCase(game, player, logger.ToUseCaseLogger());
         }
     }
-
-    internal static class UnityLoggerExtension
-    {
-        public static ILogger ToUseCaseLogger(this UnityEngine.ILogger logger)
-        {
-            return new LoaggerWrapper(logger);
-        }
-
-        private class LoaggerWrapper : Domain.UseCase.ILogger
-        {
-            readonly UnityEngine.ILogger logger;
-
-            public LoaggerWrapper(UnityEngine.ILogger logger)
-            {
-                this.logger = logger;
-            }
-
-            public void Log(string message) => logger.Log(message);
-        }
-    }
 }
