+++ Assets/Scripts/CerkeOnline/Application/UseCaseFactory/UnityLoggerExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa36a98cbc664624686687efe40b1d0d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
